requirements/dev.txt
-r base.txt
black==18.9b0
coverage==4.5.2
factory-boy==2.11.1
flake8-bugbear==18.8.0
flake8-docstrings==1.3.0
flake8-isort==2.6.0
flake8==3.7.7
ipython==7.3.0
isort==4.3.9
mock==2.0.0
pep8-naming==0.8.2
pytest==4.3.0
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-factoryboy==2.0.2
pytest-mock==1.10.4
pytest-sugar==0.9.2
